Distribute nodes across Kafka partitions more evenly
The current mechanism to assign metrics to a Kafka partition
based on the hostname.hashCode() does not generate a distribution even
enough and create some hot spots in our Kafka cluster. By calculating a
CRC32 of the hostname and using that instead we create a more even
distribution, according to our tests.