[core] normalize ffwd_tag environment variables
It is general convention that environment variables are all capitalized.

Rather than exporting an environment variable as `FFWD_TAG_foo` and knowing `foo` should
be lowercase, this change proposes normalizing all parsed tags to be lowercased. With this
`FFWD_TAG_FOO` and `FFWD_TAG_foo` will produce the same tag and be part of the same timeseries.