[agent,api] Add optional metrics for batching
Every output plugin can have an optional BatchingPluginSink in front
of it. This changes the behavior of the output, and could under high
load be a place where data is built up in queues and outstanding
writes - so it's interesting to monitor.
This commit adds some new disabled-by-default metrics for this. Enable
by specifying reportStatistics:true within the batching block for the
relevant output. Like so:

output:
  plugins:
    - type: debug
      batching:
        flushInterval: 10000
        reportStatistics: true